Retry UserInfo
