+++ store/userinfo.go
@@ -22,8 +22,11 @@ package store
 import (
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"net/url"
 	"time"
+
+	"gopkg.in/retry.v1"
 )
 
 var (
@@ -47,32 +50,48 @@ type User struct {
 
 func UserInfo(email string) (userinfo *User, err error) {
 	ssourl := fmt.Sprintf("%s/keys/%s", authURL(), url.QueryEscape(email))
+	for attempt := retry.Start(defaultRetryStrategy, nil); attempt.Next(); {
+		var resp *http.Response
+		resp, err = httpClient.Get(ssourl)
+		if err != nil {
+			if shouldRetryError(attempt, err) {
+				continue
+			}
+			break
+		}
 
-	resp, err := httpClient.Get(ssourl)
-	if err != nil {
-		return nil, err
+		if shouldRetryHttpResponse(attempt, resp) {
+			resp.Body.Close()
+			continue
+		}
+
+		defer resp.Body.Close()
+
+		switch resp.StatusCode {
+		case 200:
+			// good
+			break
+		case 404:
+			return nil, fmt.Errorf("cannot find user %q", email)
+		default:
+			return nil, respToError(resp, fmt.Sprintf("look up user %q", email))
+		}
+
+		var v keysReply
+		dec := json.NewDecoder(resp.Body)
+		if err = dec.Decode(&v); err != nil {
+			if shouldRetryError(attempt, err) {
+				continue
+			}
+			return nil, fmt.Errorf("cannot unmarshal: %s", err)
+		}
+
+		return &User{
+			Username:         v.Username,
+			SSHKeys:          v.SSHKeys,
+			OpenIDIdentifier: v.OpenIDIdentifier,
+		}, nil
 	}
-	defer resp.Body.Close()
 
-	switch resp.StatusCode {
-	case 200:
-		// good
-		break
-	case 404:
-		return nil, fmt.Errorf("cannot find user %q", email)
-	default:
-		return nil, respToError(resp, fmt.Sprintf("look up user %q", email))
-	}
-
-	var v keysReply
-	dec := json.NewDecoder(resp.Body)
-	if err := dec.Decode(&v); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal: %s", err)
-	}
-
-	return &User{
-		Username:         v.Username,
-		SSHKeys:          v.SSHKeys,
-		OpenIDIdentifier: v.OpenIDIdentifier,
-	}, nil
+	return nil, err
 }
